+++ src/test/regress/sql/float8.sql
@@ -45,7 +45,7 @@ SELECT 'Infinity'::float8 / 'Infinity'::float8;
 SELECT '42'::float8 / 'Infinity'::float8;
 SELECT 'nan'::float8 / 'nan'::float8;
 SELECT 'nan'::float8 / '0'::float8;
-SELECT 'nan'::numeric::float8;
+SELECT 'nan'::int8::float8;
 
 SELECT * FROM FLOAT8_TBL;
