+++ src/test/regress/sql/pg_lsn.sql
@@ -27,17 +27,17 @@ SELECT '0/16AE7F7' < '0/16AE7F8'::pg_lsn;
 SELECT '0/16AE7F8' > pg_lsn '0/16AE7F7';
 SELECT '0/16AE7F7'::pg_lsn - '0/16AE7F8'::pg_lsn;
 SELECT '0/16AE7F8'::pg_lsn - '0/16AE7F7'::pg_lsn;
-SELECT '0/16AE7F7'::pg_lsn + 16::numeric;
-SELECT 16::numeric + '0/16AE7F7'::pg_lsn;
-SELECT '0/16AE7F7'::pg_lsn - 16::numeric;
-SELECT 'FFFFFFFF/FFFFFFFE'::pg_lsn + 1::numeric;
-SELECT 'FFFFFFFF/FFFFFFFE'::pg_lsn + 2::numeric; -- out of range error
-SELECT '0/1'::pg_lsn - 1::numeric;
-SELECT '0/1'::pg_lsn - 2::numeric; -- out of range error
+SELECT '0/16AE7F7'::pg_lsn + 16::int8;
+SELECT 16::int8 + '0/16AE7F7'::pg_lsn;
+SELECT '0/16AE7F7'::pg_lsn - 16::int8;
+SELECT 'FFFFFFFF/FFFFFFFE'::pg_lsn + 1::int8;
+SELECT 'FFFFFFFF/FFFFFFFE'::pg_lsn + 2::int8; -- out of range error
+SELECT '0/1'::pg_lsn - 1::int8;
+SELECT '0/1'::pg_lsn - 2::int8; -- out of range error
 SELECT '0/0'::pg_lsn + ('FFFFFFFF/FFFFFFFF'::pg_lsn - '0/0'::pg_lsn);
 SELECT 'FFFFFFFF/FFFFFFFF'::pg_lsn - ('FFFFFFFF/FFFFFFFF'::pg_lsn - '0/0'::pg_lsn);
-SELECT '0/16AE7F7'::pg_lsn + 'NaN'::numeric;
-SELECT '0/16AE7F7'::pg_lsn - 'NaN'::numeric;
+SELECT '0/16AE7F7'::pg_lsn + 'NaN'::int8;
+SELECT '0/16AE7F7'::pg_lsn - 'NaN'::int8;
 
 -- Check btree and hash opclasses
 EXPLAIN (COSTS OFF)
